Implement test for logout action (successful case).
Logout is a simple GET, so there is no need to include a *functional*
test for failures.

We might consider it a failure of the logout action if it allows
unauthorized access. However, this is handled by the authenticator
defined in `controllers/Secured.java`, so any testing logic covering
this case should target the authenticator instead of the logout action.