WIP: Refactor: Use `attributes` function to dynamically set HTML attributes on root elements (el) of FeatureItemViews and ValueItemViews (instead of misusing `render` function for this purpose).
Aside from `class`, all attributes of feature and value items depend on
properties of the model instance passed in when creating (instantiating)
the corresponding View. As a consequence, they can not be set

- at the top-level of the object passed to `Backbone.View.extend()`
- inside an `attributes` *object*.

If `attributes` is a *function*, however, it is run *after* any
properties passed to a view's constructor are set, which makes it
possible to generate attributes dynamically, based on the data passed in
to the view's constructor. The view's `el` is created afterwards, based
on the views `attributes`.

---

Sources:

- http://backbonejs.org/#View-attributes
- "Setting id and className dynamically in Backbone.js views" @
  http://stackoverflow.com/q/7033134/1199226
  (relevant answer: http://stackoverflow.com/a/10337149/1199226)
- "Can't pass 'data-' attributes within View's Attributes method" @
  https://github.com/jashkenas/backbone/issues/1703