Assign random (type 4) UUID to rule nodes when creating them, and set the value of the `rule` property on all :HAS relationships that originate from a given rule to the rule's UUID.
That way, LHS of rules are not affected by rules getting renamed.
Obviously, a UUID *must not change* over the life span of the rule it is
associated with.

Also, derive UUIDs for AVM nodes from the UUIDs of the rules they are
associated with: Deriving them from rule names caused retrieval of AVMs
belonging to a specific rule to stop working after renaming that rule.