[Follow-up: a1746a0] Make sure features can be added to LHS substructures (without refreshing the page).
This fails if empty substructures of newly added complex features do not
know their UUIDs. This change set solves this problem in the following
way: When creating a new attribute-value `Pair` on the client, the value
is *not* generated and set right away. Instead, we wait for the server
to persist the model and have it include the UUID it generated for the
pair's value (=substructure) in its response; upon receiving the
response we construct the value in the same way it is constructed when
rendering the entire structure of a given rule LHS on page load.