[OutputBuilder] Add functionality for splitting strings.
When a given string is split, the first half (before the split point) is
added to the first column of the parts table of the current combination
group, and the second half is added to the second column.

Note that for event handling purposes we now wrap `outputStrings`,
`slots`, and `parts` attributes of `CombinationGroup`, `partsTable`, and
`Slot` models into `Backbone.Collection`s (instead of using plain JS
arrays).