[OutputBuilder] Make sure separators always have a border (instead of adding one on hover).
The default is for this border to be invisible; it is made visible only
when hovering over a given separator with the mouse.

By making sure separators always have a border we get rid of any
"jiggling" that used to happen when moving the mouse back and forth over
a given output string.