[OutputBuilder] Add functionality for removing individual slots from combination groups.
With respect to removing slots, parts tables behave as follows:

If they have <3 (less than three) columns, no columns can be removed.
This is because the purpose of a parts table is to allow generation of
all possible combinations of parts listed in its columns, and we can't
generate combinations from parts tables with only a single column. (As a
result, if there were only a single column left, the parts in this
column might as well be listed with the output strings. Note that it
might become necessary/desirable to provide features that support this
kind of scenario later on down the line.)

If a given table has three or more columns, *any* of these can be
removed by clicking on the "Remove" button in the header of the slot.

When removing a column that was located *between* two sibling slots, the
headers of the slots following it will be updated to reflect the fact
that in terms of their position in the table, removing said column
caused them to get shifted to the left:

1) Starting point:

   [Slot 1] [Slot 2] [Slot 3] [Slot 4]

2) Removing the 2nd slot yields:

   [Slot 1] [Slot 3] [Slot 4]

3) After system adjusts slot headers:

   [Slot 1] [Slot 2] [Slot 3]

Finally, removing a non-empty column from a table causes all parts it
contains to be removed from the table (and enclosing combination group)
as well.