[persistence] Disregard randomly generated UUID in existence check that is performed when adding a given output string to a combination group.
That is, only check the DB for existing strings with the same *content*.
This addresses the issue raised by (and mentioned in the commit message
of) 3bb0c09.

Note that unless a given output string has already been added to the
current combination group, we have no way of knowing whether it exists
in DB. Combination groups only know about the output strings they
themselves currently contain. As a result, we *have to* perform
server-side existence checks if we want to avoid adding duplicate
OutputString nodes to the DB.