Transfer JSON representation of rule RHS to client when loading OutputBuilder.
Note that in order to save some trips to the DB we currently rely on
OutputStrings and Parts returned by CombinationGroup.getStrings and
Slot.getParts methods to have both their UUIDs *and* their content set.
(Ultimately, the `getEndNodes` methods of the `HasStringRelationship`
and `HasPartRelationship` classes are responsible for adding this
information to the OutputString and Part objects they generate).