Add functionality for listing rules that are similar to a given rule in Rule Browser.
There are now several ways to bring up this listing:

- Click "=" button on entry in Rule Browser
- Click "Similar rules" button
  + on "Rule Details" page
  + in InputBuilder
  + in OutputBuilder

The set of rules that are similar to a given rule includes

- the rule itself
- any rule whose LHS incorporates the same set of features as the rule
  in question; note that this includes rules incorporating a *superset*
  of the features of the original rule

The algorithm for calculating rule similarity operates by obtaining the
set of features belonging to the LHS of the rule in question and finding
the set of all rules that incorporate each one of these features.

Note that when compiling the set of features in the first step, the
algorithm does *not* take into account the structure of the LHS of the
rule in question! Consider:

Rule A:

| X   []    |
| Y   "..." |
| Z   "..." |

Rule B:

| X   [ Z   "..." ] |
| Y   "..."         |

... where X is a complex feature, and Y and Z are atomic features.

For each of these rules, the algorithm would determine the set of
features to be {X, Y, Z}. As a result, requesting the set of similar
rules for each of these rules would yield the same set of rules (and the
result set for "Rule A" would include "Rule B" and vice versa).