[Features] Don't force user to enter new name for value after double-clicking it in the global list of values.
Before, when clicking "OK" without changing the name of the current
value, the system would display a "Name already taken" error next to the
"OK" button. There was no way to restore the original list item without
refreshing the page. Now, we simply re-render the value in question as a
list item.