WIP: Enable operations that write to the DB to be rolled back on error by wrapping them in transactions.
When removing individual targets from a given feature (cf. 1291899), try
to remove them from the DB *iff* they are values (i.e., *iff* the
"allowing" feature is atomic). Also, don't make the success of removing
a given target from an atomic feature dependent on the success of
deleting that target from the DB.