WIP: Enable operations that write to the DB to be rolled back on error by wrapping them in transactions.
Refactor: Fix class hierarchy for relationship managers w/r/t
functionality for obtaining all incoming relationships of a given node:

BaseManager <--- RelManager <--- TypedRelManager <--- AllowsManager / ...
                            ^
                             \__ UntypedRelManager

`RelManager` implements a generic version of the `to` method that is
used to obtain *all* incoming relationships of a given
node (irrespective of their type); `TypedRelManager` overrides the `to`
method and provides an implementation that is used to obtain all
incoming relationships of a given node that are of a specific type.

All remaining functionality for working with (typed) relationships has
been pushed down to `TypedRelManager`.