WIP: Enable operations that write to the DB to be rolled back on error by wrapping them in transactions.
Refactor: DRY out `CombinationGroupManager`, `RHSManager`, and
`SlotManager` by moving shared logic for deleting associated nodes to
new common superclass (`CollectionNodeManager`).

Note that even though `AVMManager` generally follows the same steps as
the managers mentioned above (w/r/t deleting AVM nodes), it could not be
refactored to use the methods provided by the `CollectionNodeManager`
class: The `disconnect` method it defines currently needs access to the
`ruleUUID` property in order to be able to disconnect the features of a
given AVM from the values they are associated with in the context of the
current rule. (The `delete` and `empty` methods, by contrast, need a
JSON representation of the AVM that does *not* include the `ruleUUID`
property in order to be able to successfully delete a given AVM node and
identify the features it points to, respectively.)