[persistence] Fix query for `startNodes` method.
This method is used to find rules embedding a given feature. When
leaving the type(s) of the relationships connecting the start
node (rule) to the end node (feature) underspecified in the query
performed by this method, the set of nodes it returns might include one
or more false positives. This is because the query will match paths
containing :ALLOWS relationships between complex features and the
features they permit in their substructures as well.