[Features] Update targets of *complex* features after renaming feature.
When selecting a complex feature A from the global feature list after
renaming a feature B that belongs to the set of acceptable targets for
A, the target list of A will display B's new name.

Prior to this commit, the page had to be refreshed to bring target lists
of complex features up-to-date.